implementation of admin dashboard. still need to work out permissioning.

+++ classes/task/cleanup_filescan_table.php
@@ -31,19 +31,27 @@ class cleanup_filescan_table extends \core\task\scheduled_task
     // if the contenthash doesnt exist in the files table then we can remove the record from the
     // block filescan table
 
+    // to get the instance id
+    // $ctx_sql = 'select ctx.instanceid from {files} f, {context} ctx where ctx.id = f.contextid';
+
+    // generate the URL based on whether the instance is a folder or mod_resource
     $sql = 'select fs.contenthash from {block_filescan_files} fs where fs.contenthash not in (select f.contenthash from {files} f)';
 
     $rs = $DB->get_recordset_sql($sql);
 
-    mtrace('Exectuing ' . $sql);
-    mtrace($rs);
+    if (!$rs->valid()) {
+      mtrace("Oops! There were no files found");
+      return false;
+    }
 
     foreach($rs as $key => $record) {
-      mtrace($key);
-      mtrace($record->contenthash);
+
+      mtrace('Deleting' . $record->contenthash);
 
       $DB->delete_records('block_filescan_files', array('contenthash' => $record->contenthash));
     }
+
+    $rs->close();
 
     mtrace('Records with missing files deleted.');
   }
